+++ rules.h
@@ -9,7 +9,7 @@ class Rules : public Turtle {
   public:
     Rules(std::string axiom, const std::map<char, std::string>& productions);
     void iterate(const unsigned& iter);
-    void interpret(const float &distance, const int &angle);
+    void interpret(const float& distance, const float& angle);
 
   private:
     std::string instructions_;
